refactor: unify yuan/quota conversion, fix price_per_second unit to yuan/s
- common/quota.go: add YuanToQuota(yuan float64) and QuotaToYuan(quota int) helpers
- model: change PricePerSecond/ModelPrices to float64 (unit: yuan/s)
- GetPricePerSecond returns float64, ModelPrices JSON uses float64
- controller: preDeductQuota = YuanToQuota(duration × price_per_second)
- service: realQuota = YuanToQuota(moneyNet)
- dto: PricePerSecond float64 in both Request and Response

+++ dto/video_channel.go
@@ -15,11 +15,11 @@ type VideoChannelRequest struct {
 	Enabled             int    `json:"enabled"`
 	Remark              string `json:"remark"`
 	SaveRequestResponse int    `json:"save_request_response"`
-	ModelMapping        string `json:"model_mapping"`
-	ModelPrices         string `json:"model_prices"`
-	PricePerSecond      int    `json:"price_per_second"`
-	PreDeductQuota      int    `json:"pre_deduct_quota"`
-	RateLimit           int    `json:"rate_limit"`
+	ModelMapping        string  `json:"model_mapping"`
+	ModelPrices         string  `json:"model_prices"`
+	PricePerSecond      float64 `json:"price_per_second"`
+	PreDeductQuota      int     `json:"pre_deduct_quota"`
+	RateLimit           int     `json:"rate_limit"`
 }
 
 // VideoChannelResponse 渠道响应（不暴露密钥）
@@ -36,13 +36,13 @@ type VideoChannelResponse struct {
 	Enabled             int    `json:"enabled"`
 	Remark              string `json:"remark"`
 	SaveRequestResponse int    `json:"save_request_response"`
-	ModelMapping        string `json:"model_mapping"`
-	ModelPrices         string `json:"model_prices"`
-	PricePerSecond      int    `json:"price_per_second"`
-	PreDeductQuota      int    `json:"pre_deduct_quota"`
-	RateLimit           int    `json:"rate_limit"`
-	CreatedAt           int64  `json:"created_at"`
-	UpdatedAt           int64  `json:"updated_at"`
+	ModelMapping        string  `json:"model_mapping"`
+	ModelPrices         string  `json:"model_prices"`
+	PricePerSecond      float64 `json:"price_per_second"`
+	PreDeductQuota      int     `json:"pre_deduct_quota"`
+	RateLimit           int     `json:"rate_limit"`
+	CreatedAt           int64   `json:"created_at"`
+	UpdatedAt           int64   `json:"updated_at"`
 }
 
 // VideoChannelStatusRequest 启用/禁用渠道
